pyproject.toml
[project]
name = "dynmap-to-pl3xmap"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml==6.0.2",
]

[dependency-groups]
dev = [
    "ruff==0.8.0",
]
